feat: complete structure of Menu component+styles
Menu component structure completed
Presentational - is pure functional components
Containers - is stateful components
Menu component render page depends on state. Change state function
passed as prop to all rendered components which allow to lift up
state to Menu component

Complete base styling. Follow 7-1 pattern
