+++ src/components/menu/container/Tutorial.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Button from '../utility/Button';
-import Title from './Title';
+import Button from '../presentational/Button';
+import Title from '../presentational/Title';
 
-class Challenges extends React.Component {
+class Tutorial extends React.Component {
   returnToMain() {
     const { returnToMain } = this.props;
     returnToMain();
@@ -12,8 +12,8 @@ class Challenges extends React.Component {
   render() {
     return (
       <div className="menu">
-        {Title('Challenges')}
-        <div className="menu__challenges">CHALLENGES</div>
+        {Title('Tutorial')}
+        <div className="menu__tutorial">TUTORIAL</div>
         <Button
           className="menu__button"
           onClick={() => this.returnToMain()}
@@ -24,8 +24,8 @@ class Challenges extends React.Component {
   }
 }
 
-Challenges.propTypes = {
+Tutorial.propTypes = {
   returnToMain: PropTypes.func.isRequired,
 };
 
-export default Challenges;
+export default Tutorial;
