+++ src/scss/base/_base.scss
@@ -6,6 +6,7 @@
 
 html {
   font-size: 10px;
+  font-family: 'Clear Sans', 'Helvetica Neue', Arial, sans-serif;
   @media (min-width: 375px) {
     font-size: 12px;
   }
