--- src/scss/components/menu/_menu.scss
@@ -1,45 +1,0 @@
-.menu {
-  width: 30rem;
-  height: 50rem;
-  @include flex(column, space-between, center);
-}
-
-.menu__title {
-  width: $menu-block-width;
-  height: $menu-title-height;
-  font-size: 4rem;
-  background-color: aqua;
-  @include flex(row, center, center);
-}
-
-.menu__button {
-  width: $menu-block-width;
-  height: $menu-btn-height;
-  font-size: 3rem;
-  cursor: pointer;
-  @include flex(row, center, center);
-}
-
-.menu__challenges {
-  width: $menu-block-width;
-  height: $menu-block-height;
-  background-color: red;
-}
-
-.menu__settings {
-  width: $menu-block-width;
-  height: $menu-block-height;
-  background-color: blue;
-}
-
-.menu__leaderboard {
-  width: $menu-block-width;
-  height: $menu-block-height;
-  background-color: purple;
-}
-
-.menu__tutorial {
-  width: $menu-block-width;
-  height: $menu-block-height;
-  background-color: green;
-}
