+++ src/scss/style.scss
@@ -5,4 +5,4 @@
 @import './base/base';
 @import './base/normalize';
 // components
-@import './components/menu/menu';
+@import './components/menu';
